PIM-5384: move hooks into dedicated class
